Add init sql file, update models
Code refactor
