+++ internal/models/room.go
@@ -1,0 +1,14 @@
+package models
+
+import (
+	"time"
+
+	"github.com/google/uuid"
+)
+
+type Room struct {
+	ID        uuid.UUID `json:"id"`
+	Name      string    `json:"name"`
+	CreatedAt time.Time `json:"createdAt"`
+	UpdatedAt time.Time `json:"updatedAt"`
+}
